renaming shared logic to projects under services folder
