+++ internal/services/projects.go
@@ -1,4 +1,4 @@
-package shared
+package services
 
 import (
 	"strconv"
